+++ src/Am.Keyward.Api/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Am.Keyward.Api;
-
-public class Class1
-{
-
-}