+++ src/Am.Keyward.Contracts/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Am.Keyward.Contracts;
-
-public class Class1
-{
-
-}